refactor: news page.tsx refactor
