+++ src/app/(route)/(community)/news/[subcategory]/page.tsx
@@ -5,7 +5,7 @@ import NewsPostItem from "../_components/NewsPostItem";
 import { useParams } from "next/navigation";
 import NewsItem from "@/app/(route)/main/_components/newsItem";
 import useSortedNewsDataList from "@/_hooks/useSortedPosts";
-import NewsTalkToolbar from "../../_components/NewsTalkToolbar";
+import NewsTalkToolbar from "../_components/NewsTalkToolbar";
 
 export default function NewsPage() {
   const params = useParams();
