+++ src/app/(route)/(community)/news/news-detail/[id]/page.tsx
@@ -14,7 +14,7 @@ import useGetNewsInfoData from "@/_hooks/useGetNewsInfoData";
 import PostNavigation from "../../../_components/PostNavigation";
 import useSortedNewsDataList from "@/_hooks/useSortedPosts";
 import { useReadNews } from "@/_hooks/useReadNews";
-import NewsTalkToolbar from "../../../_components/NewsTalkToolbar";
+import NewsTalkToolbar from "../../_components/NewsTalkToolbar";
 
 const Page = ({ params }: { params: Promise<{ id: string }> }) => {
   const [orderType, setOrderType] = useState<"DATE" | "COMMENT" | "VIEW">(
